Incremented version number to 1.1-0
Added ergm.kurtosis model that tapers the kurtosis of the terms (rather  than the variance).
This also uses a Gaussian penalty on the kurtosis being centered on 3 with a (default) s.d. of 0.6.
This provides identifiability of the kurtosis term and the model tapering.

Corrected the covariance of the ergm.tapered term from I+B to I-B (as noted by Bart Blackburn)

As a reminder, this version requires the customized "ergm" package called "ergd" 4.1-5197 or above.