Added an estimate of the bias in the coefficients introduced by tapering
Added an estimate of the bias in the coefficients introduced by tapering.

The standard ERGM coefficients have a conditional log-odds interpretation. Tapered coefficients do not have this interpretation, but are close to the conditional log-odds interpretation. This "bias" term measures the average difference between the coefficient and the conditional log-odds for the tapered ERGM model. That is, positive bias indicates the coefficient is above the conditional log-odds.